updates in rms dashboard and rms positions web part and also deleted unused webparts
