+++ src/webparts/rmsdashboardWebpart/IRmsdashboardWebpartWebPartProps.ts
@@ -1,0 +1,3 @@
+export interface IRmsdashboardWebpartWebPartProps {
+    description: string;
+  }
