+++ src/webparts/rmsdashboardWebpart/RmsdashboardWebpartWebPart.manifest.json
@@ -3,18 +3,12 @@
   "id": "0e10ff25-88e0-48e9-bef2-0d90f2017157",
   "alias": "RmsdashboardWebpartWebPart",
   "componentType": "WebPart",
-
-  // The "*" signifies that the version should be taken from the package.json
   "version": "*",
   "manifestVersion": 2,
-
-  // If true, the component can only be installed on sites where Custom Script is allowed.
-  // Components that allow authors to embed arbitrary script code should set this to true.
-  // https://support.office.com/en-us/article/Turn-scripting-capabilities-on-or-off-1f2c515f-5d7e-448a-9fd7-835da935584f
   "requiresCustomScript": false,
 
   "preconfiguredEntries": [{
-    "groupId": "5c03119e-3074-46fd-976b-c60198311f70", // Other
+    "groupId": "5c03119e-3074-46fd-976b-c60198311f70",
     "group": { "default": "Other" },
     "title": { "default": "rmsdashboard-webpart" },
     "description": { "default": "rmsdashboard-webpart description" },
