+++ src/webparts/rmsdashboardWebpart/components/IRmsdashboardWebpartProps.ts
@@ -1,3 +1,2 @@
 export interface IRmsdashboardWebpartProps {
-  description: string;
 }
