+++ src/webparts/rmsdashboardWebpart/components/IRmsdashboardWebpartState.ts
@@ -1,0 +1,3 @@
+export interface IRmsdashboardWebpartState{
+    selectedMonthName : string;
+}
