+++ src/webparts/rmsdashboardWebpart/components/RmsdashboardWebpart.module.scss
@@ -8,33 +8,17 @@
   }
 
   .row {
-    @include ms-Grid-row;
-    @include ms-fontColor-white;
-    background-color: $ms-color-themeDark;
     padding: 20px;
   }
 
-  .column {
-    @include ms-Grid-col;
-    @include ms-lg10;
-    @include ms-xl8;
-    @include ms-xlPush2;
-    @include ms-lgPush1;
+  .listItem {
+    max-width: 715px;
+    margin: 5px auto 5px auto;
+    box-shadow: 0 0 4px 0 rgba(0, 0, 0, 0.2), 0 25px 50px 0 rgba(0, 0, 0, 0.1);
   }
 
-  .title {
-    @include ms-font-xl;
-    @include ms-fontColor-white;
-  }
-
-  .subTitle {
-    @include ms-font-l;
-    @include ms-fontColor-white;
-  }
-
-  .description {
-    @include ms-font-l;
-    @include ms-fontColor-white;
+  .textfield label {
+    color:white;
   }
 
   .button {
@@ -44,17 +28,17 @@
 
     // Primary Button
     min-width: 80px;
-    background-color: $ms-color-themePrimary;
-    border-color: $ms-color-themePrimary;
-    color: $ms-color-white;
+    background-color: #0078d7;
+    border-color: #0078d7;
+    color: #ffffff;
 
     // Basic Button
     outline: transparent;
     position: relative;
     font-family: "Segoe UI WestEuropean","Segoe UI",-apple-system,BlinkMacSystemFont,Roboto,"Helvetica Neue",sans-serif;
     -webkit-font-smoothing: antialiased;
-    font-size: $ms-font-size-m;
-    font-weight: $ms-font-weight-regular;
+    font-size: 14px;
+    font-weight: 400;
     border-width: 0;
     text-align: center;
     cursor: pointer;
@@ -62,8 +46,8 @@
     padding: 0 16px;
 
     .label {
-      font-weight: $ms-font-weight-semibold;
-      font-size: $ms-font-size-m;
+      font-weight: 600;
+      font-size: 14px;
       height: 32px;
       line-height: 32px;
       margin: 0 4px;
@@ -71,4 +55,14 @@
       display: inline-block;
     }
   }
+
+  .rmsDashboardHeader{
+    float:left;
+  }
+
+  .rmsDashboardMonthSelector{
+    float: right;
+    margin-top : 20px;
+  }
+
 }
