--- src/webparts/sample/SampleWebPart.manifest.json
@@ -1,26 +1,0 @@
-{
-  "$schema": "https://dev.office.com/json-schemas/spfx/client-side-web-part-manifest.schema.json",
-  "id": "396d78d0-d1b1-415d-a608-38c673c09cbc",
-  "alias": "SampleWebPart",
-  "componentType": "WebPart",
-
-  // The "*" signifies that the version should be taken from the package.json
-  "version": "*",
-  "manifestVersion": 2,
-
-  // If true, the component can only be installed on sites where Custom Script is allowed.
-  // Components that allow authors to embed arbitrary script code should set this to true.
-  // https://support.office.com/en-us/article/Turn-scripting-capabilities-on-or-off-1f2c515f-5d7e-448a-9fd7-835da935584f
-  "requiresCustomScript": false,
-
-  "preconfiguredEntries": [{
-    "groupId": "5c03119e-3074-46fd-976b-c60198311f70", // Other
-    "group": { "default": "Other" },
-    "title": { "default": "Sample" },
-    "description": { "default": "trail and error" },
-    "officeFabricIconFontName": "Page",
-    "properties": {
-      "description": "Sample"
-    }
-  }]
-}
