--- src/webparts/sample/SampleWebPart.ts
@@ -1,56 +1,0 @@
-import * as React from 'react';
-import * as ReactDom from 'react-dom';
-import { Version } from '@microsoft/sp-core-library';
-import {
-  BaseClientSideWebPart,
-  IPropertyPaneConfiguration,
-  PropertyPaneTextField
-} from '@microsoft/sp-webpart-base';
-
-import * as strings from 'SampleWebPartStrings';
-import Sample from './components/Sample';
-import { ISampleProps } from './components/ISampleProps';
-
-export interface ISampleWebPartProps {
-  description: string;
-}
-
-export default class SampleWebPart extends BaseClientSideWebPart<ISampleWebPartProps> {
-
-  public render(): void {
-    const element: React.ReactElement<ISampleProps > = React.createElement(
-      Sample,
-      {
-        description: this.properties.description
-      }
-    );
-
-    ReactDom.render(element, this.domElement);
-  }
-
-  protected get dataVersion(): Version {
-    return Version.parse('1.0');
-  }
-
-  protected getPropertyPaneConfiguration(): IPropertyPaneConfiguration {
-    return {
-      pages: [
-        {
-          header: {
-            description: strings.PropertyPaneDescription
-          },
-          groups: [
-            {
-              groupName: strings.BasicGroupName,
-              groupFields: [
-                PropertyPaneTextField('description', {
-                  label: strings.DescriptionFieldLabel
-                })
-              ]
-            }
-          ]
-        }
-      ]
-    };
-  }
-}
