--- src/webparts/sample/components/ISampleProps.ts
@@ -1,3 +1,0 @@
-export interface ISampleProps {
-  description: string;
-}
