--- src/webparts/sample/components/ISampleState.ts
@@ -1,3 +1,0 @@
-export interface ISampleState {
-    monthValue : string;
-}
