--- src/webparts/sample/loc/mystrings.d.ts
@@ -1,10 +1,0 @@
-declare interface ISampleWebPartStrings {
-  PropertyPaneDescription: string;
-  BasicGroupName: string;
-  DescriptionFieldLabel: string;
-}
-
-declare module 'SampleWebPartStrings' {
-  const strings: ISampleWebPartStrings;
-  export = strings;
-}
