--- src/webparts/senderWp/ISenderWpWebPartProps.ts
@@ -1,3 +1,0 @@
-export interface ISenderWpWebPartProps {
-  description: string;
-}
