--- src/webparts/senderWp/SenderWpWebPart.ts
@@ -1,52 +1,0 @@
-import * as React from 'react';
-import * as ReactDom from 'react-dom';
-import { Version } from '@microsoft/sp-core-library';
-import {
-  BaseClientSideWebPart,
-  IPropertyPaneConfiguration,
-  PropertyPaneTextField
-} from '@microsoft/sp-webpart-base';
-
-import * as strings from 'senderWpStrings';
-import SenderWp from './components/SenderWp';
-import { ISenderWpProps } from './components/ISenderWpProps';
-import { ISenderWpWebPartProps } from './ISenderWpWebPartProps';
-
-export default class SenderWpWebPart extends BaseClientSideWebPart<ISenderWpWebPartProps> {
-
-  public render(): void {
-    const element: React.ReactElement<ISenderWpProps > = React.createElement(
-      SenderWp,{
-       
-      }
-    );
-
-    ReactDom.render(element, this.domElement);
-  }
-
-  protected get dataVersion(): Version {
-    return Version.parse('1.0');
-  }
-
-  protected getPropertyPaneConfiguration(): IPropertyPaneConfiguration {
-    return {
-      pages: [
-        {
-          header: {
-            description: strings.PropertyPaneDescription
-          },
-          groups: [
-            {
-              groupName: strings.BasicGroupName,
-              groupFields: [
-                PropertyPaneTextField('description', {
-                  label: strings.DescriptionFieldLabel
-                })
-              ]
-            }
-          ]
-        }
-      ]
-    };
-  }
-}
