--- src/webparts/senderWp/components/ISenderWpProps.ts
@@ -1,2 +1,0 @@
-export interface ISenderWpProps {
-}
