--- src/webparts/senderWp/components/ISenderWpState.ts
@@ -1,4 +1,0 @@
-export interface ISenderWpState {
-    text: string;
-    selectedMonthName: string;
-}
