--- src/webparts/senderWp/loc/mystrings.d.ts
@@ -1,10 +1,0 @@
-declare interface ISenderWpStrings {
-  PropertyPaneDescription: string;
-  BasicGroupName: string;
-  DescriptionFieldLabel: string;
-}
-
-declare module 'senderWpStrings' {
-  const strings: ISenderWpStrings;
-  export = strings;
-}
