--- src/webparts/senderWp/tests/SenderWp.test.ts
@@ -1,9 +1,0 @@
-/// <reference types="mocha" />
-
-import { assert } from 'chai';
-
-describe('SenderWpWebPart', () => {
-  it('should do something', () => {
-    assert.ok(true);
-  });
-});
